chore: fix compiler and clippy warnings
- Add #[allow(dead_code)] to BloomFilter methods used only in tests
- Use div_ceil() instead of manual (n + 63) / 64 calculation
- Use std::slice::from_ref() instead of cloning for single-element slice
- Fix unnecessary i64 casts in test code
- Change test value 3.14 to 3.5 to avoid approx_constant clippy error
- Use std::hint::black_box instead of deprecated criterion::black_box

+++ crates/vibesql-executor/tests/update_onepass_tests.rs
@@ -27,14 +27,14 @@ fn setup_simple_pk_table(db: &mut Database) {
     db.create_table(schema).unwrap();
 
     // Insert test data
-    for i in 1..=3 {
+    for i in 1..=3i64 {
         db.insert_row(
             "items",
             Row::new(vec![
                 SqlValue::Integer(i),
                 SqlValue::Varchar(arcstr::ArcStr::from(format!("Item {}", i))),
-                SqlValue::Integer(100 * i as i64),
-                SqlValue::Integer(10 * i as i64),
+                SqlValue::Integer(100 * i),
+                SqlValue::Integer(10 * i),
             ]),
         )
         .unwrap();
